fix(api): use Convex query API for listing moltys
The HTTP endpoint /api/moltys requires session authentication which
the CLI cannot provide with service tokens. Switch to using the public
Convex query API at .convex.cloud/api/query which allows direct query
calls without authentication.

Changes:
- Use moltys:list query via Convex query API
- Filter results client-side by ownerId
- Change CreatedAt type to float64 for Convex number format

+++ cmd/list.go
@@ -68,7 +68,7 @@ func runList(cmd *cobra.Command, args []string) error {
 
 	for _, molty := range moltys {
 		status := formatStatus(molty.Status)
-		created := formatTimestamp(molty.CreatedAt)
+		created := formatTimestamp(int64(molty.CreatedAt))
 		sandboxID := molty.SandboxID
 		if sandboxID == "" {
 			sandboxID = "-"
